refactor(skills): split source acquisition into boundary modules (S2)
Split the 594-line src/cli/skills-source.ts into six modules under
src/cli/commands/skills/source/: types (shared interfaces), selection
(pure repository-shape selection), local (local directory/Git), github
(URL/ref/API/download with GITHUB_TOKEN only at request boundary),
archive (bounded tar/process boundary), and index (install dispatcher +
updateSkill re-acquisition). The old file is deleted with no
pass-through; lifecycle.ts and the source test import from the new
directory.
